src/test/java/com/springsamples/heroesapi/services/HeroesServiceImplTest.java
package com.springsamples.heroesapi.services;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit.jupiter.SpringExtension;

import static org.assertj.core.api.Assertions.assertThat;

@ExtendWith({SpringExtension.class})
@ContextConfiguration(classes = HeroesServiceImpl.class)
class HeroesServiceImplTest {

    @Autowired
    HeroesService service;

    @BeforeEach
    void beforeEach() {

    }

    @AfterEach
    void afterEach() {

    }

    @Test
    @DisplayName("Should get hero entities from repository")
    void findAll() {
        var heroes = service.findAll();
        assertThat(heroes).isNotEmpty();
    }
}